(#148) Added json schema repository parsing (currently no schema collection support)
