+++ modules/citrus-core/src/main/java/com/consol/citrus/config/xml/SchemaParser.java
@@ -17,6 +17,7 @@
 package com.consol.citrus.config.xml;
 
 import com.consol.citrus.config.util.BeanDefinitionParserUtils;
+import com.consol.citrus.json.JsonSchema;
 import com.consol.citrus.xml.schema.WsdlXsdSchema;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
@@ -46,6 +47,9 @@ public class SchemaParser implements BeanDefinitionParser {
         } else if (location.endsWith(".xsd")){
             builder = BeanDefinitionBuilder.genericBeanDefinition(SimpleXsdSchema.class);
             BeanDefinitionParserUtils.setPropertyValue(builder, location, "xsd");
+        } else if (location.endsWith(".json")){
+            builder = BeanDefinitionBuilder.genericBeanDefinition(JsonSchema.class);
+            BeanDefinitionParserUtils.setPropertyValue(builder, location, "json");
         }
 
         if(builder != null){
