+++ modules/citrus-core/src/main/java/com/consol/citrus/config/xml/SchemaRepositoryParser.java
@@ -16,13 +16,22 @@
 
 package com.consol.citrus.config.xml;
 
+import com.consol.citrus.config.util.BeanDefinitionParserUtils;
+import com.consol.citrus.json.JsonSchemaRepository;
+import com.consol.citrus.xml.XsdSchemaRepository;
 import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.RuntimeBeanReference;
+import org.springframework.beans.factory.support.BeanDefinitionBuilder;
+import org.springframework.beans.factory.support.ManagedList;
 import org.springframework.beans.factory.xml.BeanDefinitionParser;
 import org.springframework.beans.factory.xml.ParserContext;
 import org.springframework.util.StringUtils;
+import org.springframework.util.xml.DomUtils;
 import org.w3c.dom.Element;
 
+import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 /**
  * Bean definition parser for schema-repository configuration.
@@ -32,27 +41,30 @@ import java.util.Objects;
  */
 public class SchemaRepositoryParser implements BeanDefinitionParser {
 
-    private XmlSchemaRepositoryParser xmlSchemaRepositoryParser;
-    private JsonSchemaRepositoryParser jsonSchemaRepositoryParser;
+    private static final String LOCATION = "location";
+    private static final String LOCATIONS = "locations";
+    private static final String SCHEMA = "schema";
+    private static final String SCHEMAS = "schemas";
+    private static final String ID = "id";
 
-    public SchemaRepositoryParser(){
-        xmlSchemaRepositoryParser = new XmlSchemaRepositoryParser();
-        jsonSchemaRepositoryParser = new JsonSchemaRepositoryParser();
-    }
-
-    SchemaRepositoryParser(XmlSchemaRepositoryParser xmlSchemaRepositoryParser,
-                           JsonSchemaRepositoryParser jsonSchemaRepositoryParser) {
-        this.xmlSchemaRepositoryParser = xmlSchemaRepositoryParser;
-        this.jsonSchemaRepositoryParser = jsonSchemaRepositoryParser;
-    }
+    private final SchemaParser schemaParser = new SchemaParser();
 
     @Override
     public BeanDefinition parse(Element element, ParserContext parserContext) {
+        BeanDefinitionBuilder builder = null;
+
         if(isXmlSchemaRepository(element)){
-            xmlSchemaRepositoryParser.parse(element, parserContext);
+            builder = BeanDefinitionBuilder.genericBeanDefinition(XsdSchemaRepository.class);
+            BeanDefinitionParserUtils.setPropertyReference(builder, element.getAttribute("schema-mapping-strategy"), "schemaMappingStrategy");
         }else if(isJsonSchemaRepository(element)){
-            jsonSchemaRepositoryParser.parse(element, parserContext);
+            builder = BeanDefinitionBuilder.genericBeanDefinition(JsonSchemaRepository.class);
         }
+
+        if(builder != null){
+            addLocationsToBuilder(element, builder);
+            parseSchemasElement(element, builder, parserContext);
+        }
+
 
         return null;
     }
@@ -78,5 +90,77 @@ public class SchemaRepositoryParser implements BeanDefinitionParser {
      */
     private boolean isJsonSchemaRepository(Element element) {
         return Objects.equals(element.getAttribute("type"), "json");
+    }
+
+
+    /**
+     * Adds the locations contained in the given locations element to the BeanDefinitionBuilder
+     * @param element the element containing the locations to be added to the builder
+     * @param builder the BeanDefinitionBuilder to add the locations to
+     */
+    private void addLocationsToBuilder(Element element, BeanDefinitionBuilder builder) {
+        Element locationsElement = DomUtils.getChildElementByTagName(element, LOCATIONS);
+        if (locationsElement != null) {
+            List<Element> locationElements = DomUtils.getChildElementsByTagName(locationsElement, LOCATION);
+
+            List<String> locations = locationElements.stream()
+                    .map(locationElement -> locationElement.getAttribute("path"))
+                    .collect(Collectors.toList());
+
+            if (!locations.isEmpty()) {
+                builder.addPropertyValue(LOCATIONS, locations);
+            }
+        }
+    }
+
+    /**
+     * Parses the given schema element to RuntimeBeanReference in consideration of the given context
+     * and adds them to the builder
+     * @param element The element from where the schemas will be parsed
+     * @param builder The builder to add the resulting RuntimeBeanReference to
+     * @param parserContext The context to parse the schema elements in
+     */
+    private void parseSchemasElement(Element element,
+                                     BeanDefinitionBuilder builder,
+                                     ParserContext parserContext) {
+        Element schemasElement = DomUtils.getChildElementByTagName(element, SCHEMAS);
+
+        if (schemasElement != null) {
+            List<Element> schemaElements = DomUtils.getChildElements(schemasElement);
+            ManagedList<RuntimeBeanReference> beanReferences = constructRuntimeBeanReferences(parserContext, schemaElements);
+
+            if (!beanReferences.isEmpty()) {
+                builder.addPropertyValue(SCHEMAS, beanReferences);
+            }
+        }
+
+        parserContext.getRegistry().registerBeanDefinition(element.getAttribute(ID), builder.getBeanDefinition());
+    }
+
+    /**
+     * Construct a List of RuntimeBeanReferences from the given list of schema elements under
+     * consideration of the given parser context
+     * @param parserContext The context to parse the elements in
+     * @param schemaElements The element to be parsed
+     * @return A list of RuntimeBeanReferences that have been defined in the xml document
+     */
+    private ManagedList<RuntimeBeanReference> constructRuntimeBeanReferences(
+            ParserContext parserContext,
+            List<Element> schemaElements) {
+
+        ManagedList<RuntimeBeanReference> runtimeBeanReferences = new ManagedList<>();
+
+        for (Element schemaElement : schemaElements) {
+            if (schemaElement.hasAttribute(SCHEMA)) {
+                runtimeBeanReferences.add(
+                        new RuntimeBeanReference(schemaElement.getAttribute(SCHEMA)));
+            } else  {
+                schemaParser.parse(schemaElement, parserContext);
+                runtimeBeanReferences.add(
+                        new RuntimeBeanReference(schemaElement.getAttribute(ID)));
+            }
+        }
+
+        return runtimeBeanReferences;
     }
 }
