--- modules/citrus-core/src/test/resources/com/consol/citrus/config/xml/JsonSchemaRepositoryParserTest-context.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans  xmlns="http://www.springframework.org/schema/beans"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xmlns:citrus="http://www.citrusframework.org/schema/config" 
-        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd 
-                            http://www.citrusframework.org/schema/config http://www.citrusframework.org/schema/config/citrus-config.xsd">
-
-    <citrus:schema-repository type="json" id="jsonSchemaRepository1">
-      <citrus:schemas>
-        <citrus:schema id="schema2" location="classpath:com/consol/citrus/validation/ProductsSchema.json"/>
-        <citrus:reference schema="schema1"/>
-        <citrus:reference schema="schemaCollection1"/>
-      </citrus:schemas>
-    </citrus:schema-repository>
-
-    <citrus:schema-repository type="json" id="jsonSchemaRepository2">
-      <citrus:locations>
-        <citrus:location path="classpath:com/consol/citrus/validation/*"/>
-      </citrus:locations>
-    </citrus:schema-repository>
-
-    <citrus:schema id="schema1" location="classpath:com/consol/citrus/validation/ProductsSchema.json"/>
-
-    <citrus:schema-collection id="jsonSchemaCollection1">
-      <citrus:schemas>
-        <citrus:schema location="classpath:com/consol/citrus/validation/ProductsSchema.json"/>
-      </citrus:schemas>
-    </citrus:schema-collection>
-</beans>
