+++ modules/citrus-core/src/test/resources/com/consol/citrus/config/xml/SchemaRepositoryParserTest-context.xml
@@ -41,4 +41,22 @@
     </citrus:schema-repository>
 
     <bean id="schemaMappingStrategy" class="com.consol.citrus.xml.schema.RootQNameSchemaMappingStrategy"/>
+
+    <!-- JSON -->
+    <citrus:schema id="jsonSchema1" location="classpath:com/consol/citrus/validation/ProductsSchema.json"/>
+
+    <citrus:schema-repository type="json" id="jsonSchemaRepository1">
+        <citrus:schemas>
+            <citrus:schema id="jsonSchema2" location="classpath:com/consol/citrus/validation/ProductsSchema.json"/>
+            <citrus:reference schema="jsonSchema1"/>
+            <!--<citrus:reference schema="schemaCollection1"/>-->
+        </citrus:schemas>
+    </citrus:schema-repository>
+
+    <citrus:schema-repository type="json" id="jsonSchemaRepository2">
+        <citrus:locations>
+            <citrus:location path="classpath:com/consol/citrus/validation/*"/>
+        </citrus:locations>
+    </citrus:schema-repository>
+
 </beans>
